# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sitaara Backend API
  description: >
    API for the Sitaara backend. Manually maintained — update this file
    whenever routes, request/response shapes, or error codes change in
    cmd/server/main.go or the internal/* handler packages.
  version: "0.1.0"

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: system
    description: Liveness and health checks
  - name: users
    description: User (parent) account creation and self-lookup
  - name: auth
    description: Authentication (sign-in, token issuance)
  - name: invites
    description: Invite-code generation and lookup for linking a partner
  - name: videos
    description: Video engagement (likes, watch progress) shared within a family

paths:
  /hello:
    get:
      tags: [system]
      summary: Liveness check
      description: Trivial handler used to confirm the server process is up and routing requests.
      operationId: hello
      responses:
        "200":
          description: Server is alive.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: hello
                required: [message]

  /health:
    get:
      tags: [system]
      summary: Health check
      description: Pings the database connection pool to confirm the server can reach Postgres.
      operationId: health
      responses:
        "200":
          description: Server and database are healthy.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                required: [status]
        "503":
          description: Database ping failed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: unhealthy
                required: [status]

  /users:
    post:
      tags: [users]
      summary: Create a user (parent account)
      description: |
        Creates a parent user account, per Actions.md §1 "User Creation".

        Two mutually exclusive input shapes:
        - **With `inviteCode`**: the family already linked to that invite
          code is joined. `childName`/`childDob` are ignored if present. The
          invite code is marked redeemed and the response omits `child`.
        - **Without `inviteCode`**: `childName` and `childDob` are required
          and a new family and child record are created first. The response
          includes the created `child`.

        A family may have at most 2 linked users, and no two users in the
        same family may share the same `role` (enforced by a DB unique
        constraint on `(family_id, role)` plus an application-level check).

        On success, a JWT (`token`) is returned, carrying the new user's id
        (`sub`) and `role` as claims, signed HS256, valid for 30 days.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
            examples:
              withInviteCode:
                summary: Second parent joining via invite code
                value:
                  email: dad@example.com
                  password: correcthorse
                  role: dad
                  inviteCode: ABC123XYZ
              withNewChild:
                summary: First parent, creates a new child
                value:
                  email: mom@example.com
                  password: correcthorse
                  role: mom
                  childName: Ada
                  childDob: "2023-05-01"
      responses:
        "201":
          description: User created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateUserResponse"
        "400":
          description: >
            Validation failure — malformed JSON, missing/invalid email,
            password, or role, missing childName/childDob when inviteCode
            is absent, or an invalid/future childDob. Always uses error
            code `invalid_request`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error: password must be at least 8 characters
                code: invalid_request
        "404":
          description: Invite code does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error: invite code not found
                code: invite_not_found
        "409":
          description: >
            Conflict — one of: email already registered, invite code
            already redeemed, family already has two linked parents, or
            family already has a user with the requested role.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              examples:
                emailTaken:
                  value: { error: email is already registered, code: email_already_registered }
                inviteUsed:
                  value: { error: invite code has already been redeemed, code: invite_already_used }
                twoParents:
                  value: { error: this family already has two linked parents, code: family_has_two_parents }
                roleTaken:
                  value: { error: this family already has a mom, code: family_already_has_role }
        "500":
          description: Unexpected internal/database error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error: something went wrong
                code: internal_error

  /auth/signin:
    post:
      tags: [auth]
      summary: Sign in
      description: |
        Verifies email + password and issues a fresh JWT, per Actions.md §2
        "User Sign In". The error response is identical (401
        `invalid_credentials`) whether the email doesn't exist or the
        password is wrong, so the API never reveals whether an email is
        registered.

        Not implemented from the spec (schema doesn't support them yet):
        account-status/ban checks, and last_login/last_activity tracking.
      operationId: signIn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignInRequest"
            examples:
              default:
                value:
                  email: mom@example.com
                  password: correcthorse
      responses:
        "200":
          description: Signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignInResponse"
        "400":
          description: Missing email or password. Error code `invalid_request`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error: email is required
                code: invalid_request
        "401":
          description: Email not registered, or password incorrect. Error code `invalid_credentials` in both cases.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error: invalid email or password
                code: invalid_credentials
        "500":
          description: Unexpected internal/database error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error: something went wrong
                code: internal_error

  /me:
    get:
      tags: [users]
      summary: Get the current user
      description: >
        Returns the authenticated user's own account info, resolved from the
        `sub` claim of the bearer token via the `RequireAuth` middleware
        (`internal/auth/middleware.go`).
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: Missing, malformed, or expired/invalid bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              examples:
                missing:
                  value: { error: missing or malformed authorization header, code: unauthorized }
                invalid:
                  value: { error: invalid or expired token, code: invalid_token }
        "500":
          description: Unexpected internal/database error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error: something went wrong
                code: internal_error

  /invite-codes:
    post:
      tags: [invites]
      summary: Generate an invite code
      description: >
        Mints a fresh invite code for the caller's family so a partner can
        join via `POST /users` with `inviteCode` set. This is a **rotation**:
        any existing unredeemed code for the family is deleted first, so
        previously-shared codes stop working once a new one is generated.
        Rejected once the family already has two linked parents.
      operationId: generateInviteCode
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Invite code created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteCodeResponse"
        "401":
          description: Missing, malformed, or expired/invalid bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error: missing or malformed authorization header
                code: unauthorized
        "409":
          description: The family already has two linked parents.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error: this family already has two linked parents
                code: family_has_two_parents
        "500":
          description: Unexpected internal/database error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error: something went wrong
                code: internal_error

  /invite-codes/active:
    get:
      tags: [invites]
      summary: Get the family's active invite code
      description: >
        Returns the caller's family's current unredeemed invite code, if one
        exists. "Active" means `redeemed_by_user_id IS NULL` — invite codes
        do not currently expire.
      operationId: getActiveInviteCode
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Active invite code.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteCodeResponse"
        "401":
          description: Missing, malformed, or expired/invalid bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error: missing or malformed authorization header
                code: unauthorized
        "404":
          description: No active (unredeemed) invite code exists for this family.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error: no active invite code for this family
                code: no_active_invite
        "500":
          description: Unexpected internal/database error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error: something went wrong
                code: internal_error

  /videos/{id}/partner-activity:
    get:
      tags: [videos]
      summary: Get family engagement for a video
      description: >
        Returns one entry per family member (self and partner) with their
        like state and watch progress for the given video, in a single
        query (`GetVideoActivityForFamily`). Each entry is tagged `isYou` so
        the caller's own row can be distinguished from the partner's.
      operationId: getVideoPartnerActivity
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Video id.
      responses:
        "200":
          description: Family engagement for this video.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyActivityResponse"
        "400":
          description: Malformed video id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error: invalid video id
                code: invalid_request
        "401":
          description: Missing, malformed, or expired/invalid bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              examples:
                missing:
                  value: { error: missing or malformed authorization header, code: unauthorized }
                invalid:
                  value: { error: invalid or expired token, code: invalid_token }
        "404":
          description: Video does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error: video not found
                code: video_not_found
        "500":
          description: Unexpected internal/database error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error: something went wrong
                code: internal_error

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ParentRole:
      type: string
      enum: [mom, dad, guardian]

    CreateUserRequest:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 72
          description: 72-byte max is bcrypt's hard truncation limit; longer values are rejected rather than silently truncated.
        role:
          $ref: "#/components/schemas/ParentRole"
        inviteCode:
          type: string
          description: If present, the user joins the family already linked to this code. Mutually exclusive with childName/childDob.
        childName:
          type: string
          description: Required if inviteCode is absent.
        childDob:
          type: string
          format: date
          description: Required if inviteCode is absent. Strict YYYY-MM-DD; must not be in the future.

    ChildResponse:
      type: object
      required: [id, name, dob]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        dob:
          type: string
          format: date

    CreateUserResponse:
      type: object
      required: [id, email, role, familyId, token, tokenExpiresAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/ParentRole"
        familyId:
          type: string
          format: uuid
        child:
          allOf:
            - $ref: "#/components/schemas/ChildResponse"
          description: Present only when a new child was created (i.e. no inviteCode was supplied).
        token:
          type: string
          description: >
            JWT signed HS256, valid 30 days. Claims: `sub` (user id),
            `role`, `iat`, `exp`. Store securely on the client
            (e.g. flutter_secure_storage on mobile).
        tokenExpiresAt:
          type: string
          format: date-time

    SignInRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    SignInResponse:
      type: object
      required: [id, email, role, token, tokenExpiresAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/ParentRole"
        token:
          type: string
          description: >
            JWT signed HS256, valid 30 days. Claims: `sub` (user id),
            `role`, `iat`, `exp`.
        tokenExpiresAt:
          type: string
          format: date-time

    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
        code:
          type: string
          description: Machine-readable error code for client branching.

    MeResponse:
      type: object
      required: [id, email, role, familyId, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/ParentRole"
        familyId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time

    InviteCodeResponse:
      type: object
      required: [code, familyId, createdAt]
      properties:
        code:
          type: string
          example: 7K4RXH9Q
        familyId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time

    FamilyMemberActivity:
      type: object
      required: [userId, role, isYou, liked, watchedSeconds]
      properties:
        userId:
          type: string
          format: uuid
        role:
          $ref: "#/components/schemas/ParentRole"
        isYou:
          type: boolean
        liked:
          type: boolean
        watchedSeconds:
          type: integer
        lastWatchedAt:
          type: string
          format: date-time
          nullable: true

    FamilyActivityResponse:
      type: object
      required: [videoId, members]
      properties:
        videoId:
          type: string
          format: uuid
        members:
          type: array
          items:
            $ref: "#/components/schemas/FamilyMemberActivity"